requirements.txt
numpy==1.23.5
opencv_python==4.7.0.72
positional_encodings==6.0.3
PyYAML==5.4.1
scikit_learn==1.2.2
scipy==1.9.1
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
